:tada: Add Gradle wrapper and set up initial project structure
This commit adds an initial setup for a Gradle project. It includes the Gradle wrapper with the corresponding gradlew and gradlew.bat files to ensure that the project is built and version, the repository for dependencies, and specifies that the project will use the JVM version of Kotlin. The settings.gradle.kts file configures the project name and adds the foojay plugin.

The gradle.properties file sets the Kotlin coding style to 'official'. A .gitignore file is also created to ignore unnecessary files and directories from being versioned.

These changes form the basis of any new Gradle project and were necessary to kick-start the project.