:recycle: Consolidate routine/coroutine and thread/coroutine explanations into main README
This commit updates the file structure and contents for improved readability and understanding of the concepts of Coroutine, Routine and Thread. Previously, these topics were provided in individual README files in their respective Kotlin directories. This was creating unnecessary redundancy and confusion. Now directory, making the information more accessible and understandable. Additionally, the file names for the 'main.kt'